+++ source/managers/cConfigManager.pas
@@ -10,15 +10,18 @@ type
     private
       FPresetList : TObjectList<TSettingsPreset>;
       FFileConvertPath : string;
+      FPresetNumber : Integer;
     public
       property PresetList : TObjectList<TSettingsPreset> read FPresetList write FPresetList;
       property FileConvertPath : string read FFileConvertPath write FFileConvertPath;
+      property PresetNumber : Integer read FPresetNumber write FPresetNumber;
 
       procedure SetDefault;
       procedure SaveConfigToFile;
       procedure LoadConfigFromFile;
 
       procedure SaveOnePreset(pPreset: TSettingsPreset);
+      procedure RefreshPresetNumber(pNumber: Integer);
 
       destructor Destroy; override;
       constructor Create;
@@ -70,8 +73,9 @@ begin
       //TODO: additional file data validation?
 
       valTmp := RootObject.GetValue('file_convert_path');
-      if Assigned(valTmp) then
-        FFileConvertPath := valTmp.Value;
+      if Assigned(valTmp) then FFileConvertPath := valTmp.Value;
+      valTmp := RootObject.GetValue('preset_number');
+      if Assigned(valTmp) then FPresetNumber := valTmp.AsType<Integer>;
 
       presetListTmp := RootObject.GetValue<TJSONArray>('presets');
       if Assigned(presetListTmp) then
@@ -79,14 +83,13 @@ begin
         FPresetList.Clear;
         for var I := 0 to presetListTmp.Count - 1 do
         begin
-          if I > PRESET_NUMBER then Break;
-
           tmpObject := presetListTmp.Items[I] as TJSONObject;
 
           FPresetList.Add(Serializer.Deserialize<TSettingsPreset>(tmpObject.ToJSON));
           FPresetList[I].Id := I;
         end;
       end;
+      RefreshPresetNumber(FPresetNumber); // align list to the desired number of presets
     except
       on E: Exception do
       begin
@@ -96,6 +99,37 @@ begin
   finally
     FreeAndNil(RootObject);
     Serializer.Free;
+  end;
+end;
+
+procedure TConfigManager.RefreshPresetNumber(pNumber: Integer);
+var
+  listCount: Integer;
+begin
+  if pNumber < 2 then Exit;
+  if pNumber = FPresetList.Count then Exit;
+
+  try
+    FPresetNumber := pNumber;
+    listCount := FPresetList.Count - 1;
+    if FPresetNumber > listCount then
+    begin
+      for var I := FPresetList.Count to FPresetNumber - 1 do
+      begin
+        FPresetList.Add(TSettingsPreset.Create);
+        FPresetList[I].Id := I;
+      end;
+    end
+    else
+    begin
+      for var I := listCount downto FPresetNumber do
+        FPresetList.Delete(I);
+    end;
+  except
+    on E: Exception do
+    begin
+      SaveToLog('Error while setting default values: ' + E.Message);
+    end;
   end;
 end;
 
@@ -111,6 +145,7 @@ begin
   try
     try
       RootObject.AddPair('file_convert_path', FFileConvertPath);
+      RootObject.AddPair('preset_number', FPresetNumber);
 
       tmpArray := TJSONArray.Create;
 
@@ -147,10 +182,10 @@ begin
   begin
     try
       try
-        if (pPreset.Id < 1) or (pPreset.Id > PRESET_NUMBER) then
+        if (pPreset.Id < 1) or (pPreset.Id > FPresetNumber - 1) then
           raise Exception.Create('Wrong preset id');
 
-        PresetList[pPreset.Id].AssignValues(pPreset);
+        FPresetList[pPreset.Id].AssignValues(pPreset);
         SaveConfigToFile;
       except
         on E: Exception do
@@ -168,12 +203,12 @@ procedure TConfigManager.SetDefault;
 begin
   try
     FFileConvertPath := '';
+    FPresetNumber := 4;
 
     FPresetList.Clear;
-    for var I := 0 to PRESET_NUMBER do //includes current, working settings
-    begin
-      FPresetList.Add(TSettingsPreset.Create);
-    end;
+    RefreshPresetNumber(FPresetNumber);
+//    for var I := 0 to FPresetNumber do //includes current, working settings
+//      FPresetList.Add(TSettingsPreset.Create);
   except
     on E: Exception do
     begin