+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unversioned
+
+## Added
+
+- Setting number of presets 3-10.
+
+## Changed
+
+- Layout: controls set to left.
+
 ## [1.1.0] - 2025-12-14
 
 ## Added